# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoiceCircle API",
    "description": "Backend for VoiceCircle: voice enrollment, practice scam calls, deepfake checks and Hugh companion calls.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/members/{member_id}/companion-calls": {
      "post": {
        "tags": [
          "companion"
        ],
        "summary": "Call Hugh Now",
        "operationId": "call_hugh_now_api_v1_members__member_id__companion_calls_post",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Member Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionCallOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "companion"
        ],
        "summary": "List Companion Calls",
        "operationId": "list_companion_calls_api_v1_members__member_id__companion_calls_get",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Member Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanionCallOut"
                  },
                  "title": "Response List Companion Calls Api V1 Members  Member Id  Companion Calls Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companion-calls/{call_id}": {
      "get": {
        "tags": [
          "companion"
        ],
        "summary": "Get Companion Call",
        "operationId": "get_companion_call_api_v1_companion_calls__call_id__get",
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionCallOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/members/{member_id}/companion-trends": {
      "get": {
        "tags": [
          "companion"
        ],
        "summary": "Companion Trends",
        "operationId": "companion_trends_api_v1_members__member_id__companion_trends_get",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Member Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 14,
              "title": "Days"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/circles/{circle_id}/alerts": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "List Alerts",
        "operationId": "list_alerts_api_v1_circles__circle_id__alerts_get",
        "parameters": [
          {
            "name": "circle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Circle Id"
            }
          },
          {
            "name": "include_acknowledged",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Acknowledged"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertOut"
                  },
                  "title": "Response List Alerts Api V1 Circles  Circle Id  Alerts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{alert_id}/ack": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Ack Alert",
        "operationId": "ack_alert_api_v1_alerts__alert_id__ack_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlertOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "circle_id": {
            "type": "string",
            "title": "Circle Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "acknowledged_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "circle_id",
          "type",
          "severity",
          "title",
          "message",
          "source_id",
          "acknowledged_at",
          "created_at"
        ],
        "title": "AlertOut"
      },
...
      "CompanionCallOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "circle_id": {
            "type": "string",
            "title": "Circle Id"
          },
          "senior_member_id": {
            "type": "string",
            "title": "Senior Member Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "recall_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recall Question"
          },
          "transcript": {
            "items": {
              "$ref": "#/components/schemas/TranscriptLine"
            },
            "type": "array",
            "title": "Transcript"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "flags": {
            "items": {},
            "type": "array",
            "title": "Flags"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "circle_id",
          "senior_member_id",
          "status",
          "recall_question",
          "transcript",
          "summary",
          "metrics",
          "flags",
          "duration_seconds",
          "started_at",
          "ended_at",
          "created_at"
        ],
        "title": "CompanionCallOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TranscriptLine": {
        "properties": {
          "speaker": {
            "type": "string",
            "title": "Speaker"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "t_ms": {
            "type": "integer",
            "title": "T Ms",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "speaker",
          "text"
        ],
        "title": "TranscriptLine"
      },
      "TrendPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "wpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wpm"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "filler_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filler Rate"
          },
          "recall": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recall"
          },
          "mood": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "TrendPoint"
      },
      "TrendsOut": {
        "properties": {
          "member_id": {
            "type": "string",
            "title": "Member Id"
          },
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "baseline": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Baseline"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flags"
          }
        },
        "type": "object",
        "required": [
          "member_id",
          "days",
          "points",
          "baseline",
          "flags"
        ],
        "title": "TrendsOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}